Spacing and commenting
I reformated it again! :(, better spacing now, before i was trying to
see how compact i could get it and see if it was usable. Nope. so now
it's MUCH easier to read, and added lots new comments to tell you how
the program works, and also changed the way how the snoring in
Game5(sleeping) is printed. for loop. so. better coments, better
spacing, :)