Major remodeling of structure.
instead of functions on the same level, we now have nested classes and functions for each path of story possibilities. to make it so it doesn't nest to deep. there are blocks of major story moments. using class and function naming depending on whether they start with a '_', ActionMenu can now be all automated instead of typing all that shit out like before!. much smoooooother.