predate_targets function overhaul. also some exception statement changes
predate_targets will now add all of the predated mobs skills to your own, ik it's not exactly canon, but it's easier for coding and for the player so they don't have to use mimic all the time. They'll still have to use mimic for some storyline stuff down the road however.