Reorganizing, and changing story structure
re doing the structure of pathways and actions player can take, to better suite the new feature to hide available actions, but still there to turn back on if lost. Note, even with the action menu on (by default is off) it won't show which action will progress the story.