seems to be working now
no need for loop variable mechanic, where before it would check if the current action that was matched with user input, started with a _ and if so it'll change the loop boolean var. Now all functions (actions, paths, whatever u call it) will loop until a new action_menu is called.
Also now in story when you eat (predate) water it'll give u the new skill but won't show that message until you use it to try to escape.
set function (action) name to _hfunc (hide function) to hide the function from menu, and set it's __subs list so user can use it