text_crawl > textcrawl, show_ascii > show_art
renamed functions that has to do with mobs and games, and some other ones.
Also the great sage ascii art will load with textcrawl effect unless you start game with -a, which won't exactly be possible when made into executable (maybe, idk). And after the ascii art is when you are prompted to enable/disable textcrawl effect, so basically the first ascii art (the great sage one) will always load with textcrawl effect