merging beginning part from chapter2 to chapter one
this is so when leaving the cave is a new chapter, and everything inside will be chapter1.
Updating the was part chapter2 bits to new system with __subs and classes.
Also added cave_actions class, which can be added to classes to add extra actions so you don't have to keep adding them, like eat_grass and eat_ore for chapter1. However, if you do want to update them inline, you'll have to update the __subs too, which you'll have to do something like __subs = cave_actions.eat_grass.__subs