asks for enabling show_actions instead of hardcore
Now asks user if they want to see playable actions or not instead of just enabling hardcore mode, added a line to start_banner if player wants to enable hardcore. 
Changed show_hud to show_actions since enabling or disabling this won't effect the showing of current mimicry and targets.