Adds action name to __subs with hfunc actions
Without this line, to be able to activate hfunc actions (including _hfunc), you would need to add the action name to it's __subs also, e.g. if you have hfunc_attack action, without __subs list, you won't be able to activate it, no matter what the player types. Unless you 'attack' to __subs. With this line you don't have to, you can activate the action like any other as if it didn't have hfunc in it's name. hfunc is just there (was suppose to be hidden function) to hide it from the HUD if that was enabled. 
_hfunc_craft_full_potion action in tensei_1 line 93 was effectively unreachable because it doesn't have a __subs. Now it doesn't need it.