Fixed bugs when player uses commands on game objects that shouldn't be
Fixed bugs if user used a game command on something that doesn't support it, e.g. using 'stats' on a game item or attribute instead of character. 
Also added checkers for adding and removing items in inventory to check if it's actually a game item, e.g. so you can't do 'remove water blade', etc.